+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+- Refactor deployment errors parsing: use JSON output instead of text output
 - [hardis:org:test:apex](https://sfdx-hardis.cloudity.com/hardis/org/test/apex/): Display the number of failed tests in messages and notifications
 - Obfuscate some data from text log files
+- Kill some exit handlers in case they are making the app crash after a throw SfError
 
 ## [5.10.1] 2024-12-12